tests/wpt/tests/css/css-animations/parsing/animation-iteration-count-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: getComputedStyle().animationIterationCount</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-iteration-count">
<meta name="assert" content="animation-iteration-count computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #container {
    container-type: inline-size;
    width: 100px;
  }
</style>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<script>
test_computed_value("animation-iteration-count", "0, infinite, 3");
test_computed_value("animation-iteration-count", "calc(10 + (sign(2cqw - 10px) * 5))", "5");
</script>
</body>
</html>
